backend/requirements.txt
# LangChain + LangGraph 多智能体框架
langchain==1.2.15
langchain-openai>=1.1.0
langgraph==1.1.8
langchain-core>=1.2.0

# MCP协议
mcp>=1.0.0
langchain-mcp-adapters>=0.2.0

# SSE流式输出
sse-starlette>=1.6.0

# FastAPI和相关依赖
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# HTTP客户端
httpx>=0.27.0
aiohttp>=3.10.0

# 环境变量管理
python-dotenv>=1.0.0

# CORS支持
python-multipart>=0.0.9

# 日志
loguru>=0.7.0

# 高德地图服务
requests>=2.31.0

# 图片服务 (Unsplash)
Pillow>=10.0.0

# Redis
redis>=5.0.0

# 语义检索
sentence-transformers>=2.2.0

# 其他工具
python-dateutil>=2.8.2
nest-asyncio>=1.5.0
